# cm.klg.media/specs/openapi/inbound/main.yml
---
openapi: "3.0.3"
info:
  title: "MEDIA MANAGEMENT REST API"
  description: "API for managing media fichier"
  version: "0.0.1-SNAPSHOT"
paths:
  /documents:
    post:
      tags:
      - "Document"
      security:
      - bearerAuth: []
      summary: "Create a document"
      description: "Upload a document and store its metadata"
      operationId: "createDocument"
      parameters:
      - name: "type"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/DocumentType"
      - name: "visibility"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/DocumentVisibility"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: "object"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "201":
          description: "Document created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponse"
        "400":
          description: "MEDIA_400_001 | MEDIA_400_002 | MEDIA_400_003"
        "500":
          description: "MEDIA_500_001 | MEDIA_500_002"
    get:
      tags:
      - "Document"
      security:
      - bearerAuth: []
      summary: "Get all documents"
      description: "Retrieve all document metadata"
      operationId: "getAllDocuments"
      responses:
        "200":
          description: "Documents successfully retrieved"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Document"
        "500":
          description: "MEDIA_500_003"
  /documents/{documentId}:
    get:
      tags:
      - "Document"
      security:
      - bearerAuth: []
      summary: "Get document with content"
      description: "Retrieve document metadata and actual file content (base64-encoded)"
      operationId: "getDocumentById"
      parameters:
      - name: "documentId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Document found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: "MEDIA_404_001"
        "500":
          description: "MEDIA_500_003"
    delete:
      tags:
      - "Document"
      security:
      - bearerAuth: []
      summary: "Mark document for deletion"
      description: "Soft delete document. Actual deletion is performed later by a\
        \ scheduler."
      operationId: "markAsDelete"
      parameters:
      - name: "documentId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "204":
          description: "Document marked for deletion"
        "404":
          description: "MEDIA_404_001"
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
    Document:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        isPublic:
          type: "boolean"
        type:
          $ref: "#/components/schemas/DocumentType"
        createdAt:
          type: "string"
          format: "date-time"
        name:
          type: "string"
        size:
          type: "number"
        content:
          type: "string"
          format: "byte"
    CreationResponse:
      type: "object"
      required:
      - "newId"
      properties:
        newId:
          type: "string"
          format: "uuid"
    DocumentType:
      type: "string"
      enum:
      - "IMAGE_JPEG"
      - "IMAGE_JPG"
      - "IMAGE_PNG"
    DocumentVisibility:
      type: "string"
      enum:
      - "PUBLIC"
      - "PRIVATE"